Moved comment mutation and form state to the comment form component
